// tracks-front/src/entities/vehicle/model/types.ts
import {User} from "../../user/model/types";

export type Vehicle = {
    id: string;
    digits: string;
    type: string;
    modelYear: string;
    carryingCapacity: string;
    fuelType: string;
    gabarites: string;
    cargoDimensions: string;
    passengersCapacity: string;
    photoUrl: string;
}